Track Cargo.lock to allow deterministic builds
The sentiment around version controlling Cargo.lock seems to
differ/evolve a bit. The relevant SO answer quotes a bit dated answer
from Rust FAQ [1] that libraries should not version control Cargo.lock.
However, the current [2] answer rather makes the sentiment that there
are cases, where the lock file is not respected. I would say that this
is the burden of the user and actually being explicit about what
Cargo.lock was succesfully used to build has its merit. It's,
for example, needed also for deterministic builds that I am working on.

[1] https://stackoverflow.com/a/63967863
[2] https://doc.rust-lang.org/cargo/faq.html#why-have-cargolock-in-version-control